Improvement node editor window

+++ Assets/Scripts/AI/NodeEditor.cs
@@ -44,9 +44,10 @@ namespace DoubleMMPrjc
 
             public void OnGUI()
             {
-                GUILayout.BeginHorizontal( GUILayout.ExpandWidth( true ) );    // 1 - START
-                GUILayout.BeginVertical( "Box", GUILayout.MaxWidth( 200 ), GUILayout.MaxHeight( 300 ), GUILayout.MinHeight( 300 ), GUILayout.ExpandWidth( false ) );    // 2.1 - START
-                GUILayout.BeginHorizontal();    // 3.1 - START
+                GUILayout.BeginHorizontal( GUILayout.ExpandWidth( true ) );    // 1 - START (H)
+                GUILayout.BeginVertical( "Box", GUILayout.MaxWidth( 200 ), GUILayout.MaxHeight( 300 ), GUILayout.MinHeight( 300 ), GUILayout.ExpandWidth( false ) );    // 2 - START (V)
+                GUILayout.BeginHorizontal();    // 3 - START (H)
+
                 if (GUILayout.Button( "New", GUILayout.Width( 50 ) )) {
                     nodeNeighborhoods.Add( Create( SceneManager.GetActiveScene().name, data ) );
                 }
@@ -59,31 +60,66 @@ namespace DoubleMMPrjc
                 if (GUILayout.Button( "Del", GUILayout.Width( 50 ) )) {
                     DeleteNeighboor( SceneManager.GetActiveScene().name, selectedNodeNeighborhood );
                 }
-                GUILayout.EndHorizontal();  // 3.1 - END
+                GUILayout.EndHorizontal();  // 3 - END (H)
 
-                if (nodeNeighborhoods.Count > 0) {
-                    nodeListScroll = GUILayout.BeginScrollView( nodeListScroll, false, false, GUILayout.ExpandWidth( true ) );  // 3.2 - START
-                    selectedIndex = GUILayout.SelectionGrid( selectedIndex, Utility.ToExtendedStringArray( nodeNeighborhoods.ToArray() ), 1, GUILayout.Width( 200 ), GUILayout.ExpandWidth( true ) );
-                    GUILayout.EndScrollView();  // 3.2 - END
-                } else {
-                    EditorGUILayout.HelpBox( "Create node and connection clicling on New button", MessageType.Warning );
-                }
+                ShowNeighborhoodList();
 
-                GUILayout.EndVertical();    // 2.1 - END
+                GUILayout.EndVertical();    // 2 - END (V)
+
                 GUILayout.Space( 25 );
 
                 if (editedNodeNeighborhood != null) {
-                    GUILayout.BeginVertical();  // 2.2 - START
+                    ShowEditLayout();
+                } else {
+                    EditorGUILayout.HelpBox( "Select node and click edit button to start editing node", MessageType.Info );
+                }
 
-                    GUILayout.Label( "Edited node: " + editedNodeNeighborhood.ToExtendedString(), EditorStyles.boldLabel, GUILayout.ExpandWidth( false ) );
+                GUILayout.Space( 25 );
 
+                if (GUI.changed) {
+                    EditorUtility.SetDirty( data );
+                }
+                GUILayout.EndHorizontal();  // 1 - END (H)
+            }
+
+            private void ShowNeighborhoodList()
+            {
+                if (nodeNeighborhoods.Count > 0) {
+                    nodeListScroll = GUILayout.BeginScrollView( nodeListScroll, false, false, GUILayout.ExpandWidth( true ) );  // 1 - START (SV)
+                    selectedIndex = GUILayout.SelectionGrid( selectedIndex, Utility.ToExtendedStringArray( nodeNeighborhoods.ToArray() ), 1, GUILayout.Width( 200 ), GUILayout.ExpandWidth( true ) );
+                    SelectNeighborhood( selectedIndex );
+                    GUILayout.EndScrollView();  // 1 - END (SV)
+                } else {
+                    EditorGUILayout.HelpBox( "Create node and connection clicling on New button", MessageType.Warning );
+                }
+            }
+
+            private void ShowEditLayout()
+            {
+                GUILayout.BeginVertical();  // 1 - START
+
+                GUILayout.Label( "Edited node: " + editedNodeNeighborhood.ToExtendedString(), EditorStyles.boldLabel, GUILayout.ExpandWidth( false ) );
+
+                GUILayout.BeginHorizontal(); // 2 - START
+
+                if (GUILayout.Button( "Focus", GUILayout.ExpandWidth( false ) )) {
+                    FocusNodeOnScene( editedNodeNeighborhood );
+                }
+
+                if (GUILayout.Button( "Delete", GUILayout.ExpandWidth( false ) )) {
+                    DeleteNeighboor( SceneManager.GetActiveScene().name, editedNodeNeighborhood, true );
+                }
+
+                GUILayout.EndHorizontal(); // 2 - END
+
+                if (editedNodeNeighborhood != null) {
                     neihgborsListScroll = GUILayout.BeginScrollView( neihgborsListScroll, false, false,
-                                                                    GUILayout.ExpandWidth( false ) ); // 3 - START
+                                                                    GUILayout.ExpandWidth( false ) ); // 2 - START
                     Edge toRemove = null;
 
                     // CREATING LIST OF EDGES CONNECTING WITH THIS NODE
                     foreach (Edge edge in editedNodeNeighborhood.Edges) {
-                        GUILayout.BeginHorizontal(); // 4 - START
+                        GUILayout.BeginHorizontal(); // 3 - START
 
                         GUILayout.Label( edge.ToString(), GUILayout.ExpandWidth( true ), GUILayout.Width( 140 ) );
 
@@ -95,34 +131,60 @@ namespace DoubleMMPrjc
                             toRemove = edge;
                         }
 
-                        GUILayout.EndHorizontal(); // 4 - END
+                        GUILayout.EndHorizontal(); // 3 - END
                         GUILayout.Space( 5 );
                     }
-                    GUILayout.EndScrollView();  // 3 - END
+
+                    GUILayout.EndScrollView();  // 2 - END
+
+
+
+                    bool disableButton = ( selectedNodeNeighborhood == null ) || ( selectedNodeNeighborhood == editedNodeNeighborhood );
+                    if (disableButton) {
+                        EditorGUILayout.HelpBox( "Select another node to add", MessageType.Info );
+                    } else if (editedNodeNeighborhood.Contains( selectedNodeNeighborhood.NodeId )) {
+                        EditorGUILayout.HelpBox( "Edited node has already selected node, select another to add", MessageType.Info );
+                        disableButton = true;
+                    }
+
+                    EditorGUI.BeginDisabledGroup( disableButton );
 
                     if (GUILayout.Button( "Add selected node" )) {
-                        if (selectedNodeNeighborhood != null) {
-                            if (selectedNodeNeighborhood != editedNodeNeighborhood) {
-                                try {
-                                    Edge edge = new Edge( editedNodeNeighborhood.NodeId, selectedNodeNeighborhood.NodeId, Direction.BOTH );
-                                    editedNodeNeighborhood.AddEdge( edge );
-                                    selectedNodeNeighborhood.AddEdge( edge );
-                                } catch (System.Exception e) {
-                                    Debug.LogError( e.Message );
+                        if (selectedNodeNeighborhood != editedNodeNeighborhood) {
+                            try {
+                                Edge _edge = new Edge( editedNodeNeighborhood.NodeId, selectedNodeNeighborhood.NodeId, Direction.BOTH );
+                                editedNodeNeighborhood.AddEdge( _edge );
+                                selectedNodeNeighborhood.AddEdge( _edge );
+                                if (GUI.changed) {
+                                    EditorUtility.SetDirty( selectedNodeNeighborhood );
                                 }
+                            } catch (System.Exception e) {
+                                Debug.LogError( e.Message );
                             }
                         }
                         editedNodeNeighborhood.RemoveEdge( toRemove );
-
-                        GUILayout.Space( 15 );
-                        GUILayout.EndVertical();    // 2.2 - END
                     }
-                    GUILayout.Space( 25 );
-                    GUILayout.EndHorizontal();  // 1 - END               
-                } else {
-                    EditorGUILayout.HelpBox( "Select node and click edit button to start editing node", MessageType.Info );
+
+                    EditorGUI.EndDisabledGroup();
+
+                    GUILayout.Space( 15 );
+                    if (GUI.changed) {
+                        EditorUtility.SetDirty( editedNodeNeighborhood );
+                    }
                 }
-                SelectNeighborhood( selectedIndex );
+                GUILayout.EndVertical();    // 1 - END
+            }
+
+            /// <summary>
+            /// Selects asset from list
+            /// </summary>
+            /// <param name="index"></param>
+            private void SelectNeighborhood(int index)
+            {
+                if (index >= nodeNeighborhoods.Count)
+                    selectedNodeNeighborhood = null;
+                else
+                    selectedNodeNeighborhood = nodeNeighborhoods[index];
             }
 
             public void ReloadNodeConnections(string sceneName)
@@ -145,18 +207,6 @@ namespace DoubleMMPrjc
             }
 
             /// <summary>
-            /// Selects asset from list
-            /// </summary>
-            /// <param name="index"></param>
-            public void SelectNeighborhood(int index)
-            {
-                if (index >= nodeNeighborhoods.Count)
-                    selectedNodeNeighborhood = null;
-                else
-                    selectedNodeNeighborhood = nodeNeighborhoods[index];
-            }
-
-            /// <summary>
             /// Focuses node on scene that belongs to asset object from list
             /// </summary>
             /// <param name="selectedNeighborhood"></param>
@@ -172,16 +222,47 @@ namespace DoubleMMPrjc
                 }
             }
 
-            public void DeleteNeighboor(string sceneName, NodeNeighborhood neighborhood)
+            public void DeleteNeighboor(string sceneName, NodeNeighborhood neighborhood, bool confirmation = false)
             {
-                Node node = GetNodeFromScene( neighborhood.NodeId );
-                if (node != null) {
-                    DestroyImmediate( node.gameObject );
-                } else {
-                    Debug.LogError( "Node with id: " + neighborhood.NodeId + " dosn't exists on scene" );
+                bool delete = true;
+                if (confirmation) {
+                    delete = EditorUtility.DisplayDialog( "Node delete", "Are u sure to delete " +
+                                                        neighborhood.ToExtendedString() + "? It will remove also node " +
+                                                        GetNodeFromScene( neighborhood.NodeId ).ToExtendedString() + "?",
+                                                        "Ok", "Cancel" );
                 }
-                nodeNeighborhoods.Remove( neighborhood );
-                AssetDatabase.DeleteAsset( "Assets/Resources/" + NEIGHBOORS_FOLDER + "/" + sceneName + "/" + neighborhood.ToString() + ".asset" );
+                if (delete) {
+                    Node node = GetNodeFromScene( neighborhood.NodeId );
+                    if (node != null) {
+                        DestroyImmediate( node.gameObject );
+                    } else {
+                        Debug.LogError( "Node with id: " + neighborhood.NodeId + " dosn't exists on scene" );
+                    }
+                    nodeNeighborhoods.Remove( neighborhood );
+                    RemoveConnections( neighborhood );
+                    AssetDatabase.DeleteAsset( "Assets/Resources/" + NEIGHBOORS_FOLDER + "/" + sceneName + "/" + neighborhood.ToString() + ".asset" );
+                    editedNodeNeighborhood = null;
+                }
+            }
+
+            public void RemoveConnections(NodeNeighborhood neighborhood)
+            {
+                foreach (Edge edge in neighborhood.Edges) {
+                    NodeNeighborhood other = GetNeighborhood( edge.GetAnotherId( neighborhood.NodeId ) );
+                    if (other != null) {
+                        other.RemoveEdge( edge );
+                    }
+                }
+                neighborhood.Edges.Clear();
+            }
+
+            public NodeNeighborhood GetNeighborhood(long id)
+            {
+                foreach (NodeNeighborhood neighborhood in nodeNeighborhoods) {
+                    if (neighborhood.NodeId == id)
+                        return neighborhood;
+                }
+                return null;
             }
 
             public static NodeNeighborhood[] LoadNodeConnnections(string sceneName)
@@ -192,7 +273,7 @@ namespace DoubleMMPrjc
 
             public static string GenerateName(string startName, long id)
             {
-                return startName  + id;
+                return startName + id;
             }
 
             public static NodeNeighborhood Create(string sceneName, SceneNodeData data)
